instagram-scraper/requirements.txt
scrapling[all]
patchright
supabase>=2.0.0
python-dotenv>=1.0.0
